# Make di more efficient