use cache and goroutines to evaluate built-in prop native codes faster