@Syuparn 2020-11-07
benchmarks:

before:

```
goos: windows
goarch: amd64
pkg: github.com/Syuparn/pangaea/runscript
BenchmarkSetup-4   	       1	1529908700 ns/op	530447760 B/op	  418335 allocs/op
PASS
ok  	github.com/Syuparn/pangaea/runscript	1.580s
```

after caching:

```
goos: windows
goarch: amd64
pkg: github.com/Syuparn/pangaea/runscript
BenchmarkSetup-4   	       3	 337763833 ns/op	120349458 B/op	  100071 allocs/op
PASS
ok  	github.com/Syuparn/pangaea/runscript	2.077s
```

after caching & concurrent evaluating

```
goos: windows
goarch: amd64
pkg: github.com/Syuparn/pangaea/runscript
BenchmarkSetup-4   	       5	 246320980 ns/op	120326329 B/op	  100036 allocs/op
PASS
ok  	github.com/Syuparn/pangaea/runscript	2.512s
```

6 times faster 😆